代理搜索用户

+++ micro-common-service/src/main/java/cn/gaozheng/sales/service/MemberService.java
@@ -15,6 +15,6 @@ public interface MemberService {
 
     List<UserInfo> memberInfoNotIncludeDelegate(MemberListParm req);
 
-    UserInfo memberInfo(Integer userId);
+    UserInfo memberInfo(Long userId);
 
 }
